Short marker rotation
In this code, I added what I mentioned about angle rotation last time in wp. Please implement it in the main code. Changes are in the setObjCardinality method and from lines 2607 to 2661 I added a new function. For this function I did not go details, it just works and does what I want, so I did not comment on the function.